Course Scheduling Project
This system is needed to help colleges when students are registering
for classes. Its intention is to present available classes and times to
which the student can add to their future semester schedule. It is
beneficial to the college institutions since it eliminates the need for
written documents and the process is done quicker and the students will
be aware of the classes offered quickly and easily. This earns the
institution profits from classes enrolled in. The Requirements Document
is crucial to this project since it lists all the necessary information
needed to complete it. This project is intended to be versatile but
small and effective so it is adaptable and marketable. Compare with
CunyFirst, this system can be more easily arrange student schedule.